fix(dashboard): replace self-calling restart logic with standalone endpoint
