+++ rustmail/src/api/handler/bot/restart.rs
@@ -11,46 +11,31 @@ use tokio::sync::Mutex;
 pub async fn handle_restart_bot(
     State(bot_state): State<Arc<Mutex<BotState>>>,
 ) -> (StatusCode, Json<&'static str>) {
-    let token = {
-        let state_lock = bot_state.lock().await;
-        state_lock.internal_token.clone()
-    };
-
     let state_lock = bot_state.lock().await;
 
     match state_lock.status {
         BotStatus::Stopped => {
             drop(state_lock);
 
-            let result = ping_internal("/api/bot/start", &token).await.map_err(|_| {
-                (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json("Failed to restart bot"),
-                )
-            });
-
-            println!("Restart result: {:?}", result);
-
-            (StatusCode::OK, Json("Bot is starting"))
+            match start_bot(bot_state).await {
+                StartBotResponse::Success(status, message) => (status, message),
+                StartBotResponse::Conflict(status, message) => (status, message),
+            }
         }
         BotStatus::Running { .. } => {
             drop(state_lock);
 
-            let _ = ping_internal("/api/bot/stop", &token).await.map_err(|_| {
-                (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json("Failed to stop bot"),
-                )
-            });
+            match stop_bot(bot_state.clone()).await {
+                StopBotResponse::Success(..) => {}
+                StopBotResponse::Conflict(status, message) => return (status, message),
+            }
 
             sleep(Duration::from_secs(2)).await;
 
-            let _ = ping_internal("/api/bot/start", &token).await.map_err(|_| {
-                (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json("Failed to stop bot"),
-                )
-            });
+            match start_bot(bot_state).await {
+                StartBotResponse::Success(..) => {}
+                StartBotResponse::Conflict(status, message) => return (status, message),
+            }
 
             (StatusCode::OK, Json("Bot is restarting"))
         }
