+++ rustmail/src/api/utils/mod.rs
@@ -1,9 +1,9 @@
 pub mod get_user_id_from_session;
 pub mod permissions_cache;
-pub mod ping_internal;
 pub mod user_permissions;
+pub mod bot;
 
 pub use get_user_id_from_session::*;
 pub use permissions_cache::*;
-pub use ping_internal::*;
 pub use user_permissions::*;
+pub use bot::*;
